test.py: move 'build' directory creation to TestCppArray.setUp()
As a first step in discovering how to use super(), delegate the creation
of the build directory to TestCppArray.setUp() and call the latter using
super() in TestCppArray1D.setUp().

Additionnally, since argparse happily accepts '-vv' for unittest's '-v'
argument (which is declared with the 'store_const' action [1]), use this
to increase verbosity and print the make output when unittest is ran
with '-vv'.

[1] https://github.com/python/cpython/blob/70f8ebe503fc7748db9b0c51cfbba20b1846a3c8/Lib/unittest/main.py#L169-L171